summarizer/requirements.txt
dagster
dagster-postgres
psycopg2-binary
requests
openai
python-dotenv
langgraph
langchain
langchain_openai
langchain-google-genai
tavily-python
sentence-transformers
